pyproject.toml
[tool.poetry]
name = "luckyboxbot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.109.2"
aiogram = "^3.4.1"
uvicorn = "^0.27.1"
tortoise-orm = "^0.20.0"
python-dotenv = "^1.0.1"
jinja2 = "^3.1.3"
pydantic-settings = "^2.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
